backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
apscheduler==3.11.0
PyYAML==6.0.2
pypdf==5.9.0
reportlab==4.4.3
pytest==8.4.1
pytest-asyncio==1.1.0

